requirements.txt
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
scikit-learn==1.4.2
streamlit==1.32.0
